requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
mlflow==2.7.1
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pyyaml==6.0.1
pytest==7.4.2
evidently==0.4.5